examples/restfulpy_demo/requirements.txt
restfulpy
sqlalchemy-media
pytest-pudb

